Omaha #3477 changed plot info locations to floats

Former-commit-id: 3b97fd15a7a3f5d293859ed086844af660ee9426 [formerly f8a1e79288acbe72e876c449f3ef764397c0743d] [formerly 8738e9cecec47e090abd6822026187e958b357f7 [formerly 00504d6aa7fefd7b6d28701aae777c0fb7d5ace5]]
Former-commit-id: 8738e9cecec47e090abd6822026187e958b357f7
Former-commit-id: a180c212c921a6a0bbeb490045c05185d2207ef3

+++ ncep/gov.noaa.nws.ncep.viz.rsc.plotdata/src/gov/noaa/nws/ncep/viz/rsc/plotdata/plotModels/NcPlotImageCreator.java
@@ -72,6 +72,7 @@ import com.vividsolutions.jts.geom.Coordinate;
  * ------------ ---------- ----------- --------------------------
  * 05/20/2013     988        Archana.S    Initial creation.
  * 02/26/2014    1061        B. Hebbard   Relax tolerance for extent/zoom compare to avoid infinite loop.
+ * Aug 08, 2014  3477        bclement     changed plot info locations to floats
  */
 
 public class NcPlotImageCreator {
@@ -892,7 +893,7 @@ public class NcPlotImageCreator {
          * @param lon
          * @return
          */
-        private String createStationMapKey(Double lon, Double lat) {
+        private String createStationMapKey(Float lon, Float lat) {
             return new String("" + Math.round(lon * 1000.0) + ","
                     + Math.round(lat * 1000.0));
